% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{clean_data}
\alias{clean_data}
\title{Clean Data}
\usage{
clean_data(data, week, year, region)
}
\description{
This function cleans data based on given parameters to make it suitable for
graphical output produced by this package
}
