% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCBIsearch.R
\name{ncbi_search}
\alias{ncbi_search}
\title{NCBI Search}
\usage{
ncbi_search(search_term, db, related = TRUE)
}
\arguments{
\item{search_term}{The term you would like to do a database search for}

\item{db}{The database you would like to use for the search}

\item{related}{Boolean value, deafaults to TRUE, set to false if you do not
want to view all the related link ids}
}
\value{
The ids for the linked records from NCBI and ids for the linked
records with related information from other databases
}
\description{
This function allows you to input a search term and returns the databases and
related id's for the records of the search and any linked records for the
search
}
\examples{
ncbi_search("AH1", "assembly")
}
