% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotData.R
\name{plot_data}
\alias{plot_data}
\title{Generates data plots}
\usage{
plot_data(type, week, year, region)
}
\arguments{
\item{type}{The type of data you would like to generate the plot for. Options
are immunization data (imm) and influenza case data for Canada (CA_flu,
CA_strains)}
}
\value{
Generated plot
}
\description{
This function generates plots for Canada Immunization rates in 2020,
Influenza activity in Canada from August to December 2022, and Influenza
Confirmed Cases from January 2020 to December 2022.
}
\examples{
plot_data("imm")
}
