% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFlulytics.R
\name{runFlulytics}
\alias{runFlulytics}
\title{A Shiny App for NCBI search}
\usage{
runFlulytics()
}
\value{
Starts shiny app
}
\description{
This function starts a shiny app where the user can choose which parameters
they would like to conduct an ncbi search with. The results are outputted for
viewing in the console, and are formatted as ncbi record ids.
}
\examples{
runFlulytics()
}
