% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewData.R
\name{view_data_table}
\alias{view_data_table}
\title{Gets Data Tables}
\usage{
view_data_table(data_type)
}
\arguments{
\item{data_type}{A character string of desired data to get table for}
}
\value{
The data in a data frame, opens up separate window to view the
original data files within the GUI
}
\description{
This function produces allows the user to view the data provided in this
package
}
\examples{
view_data_table("CA_Imm")
}
